requirements.txt
flask
twilio
pyngrok
google-generativeai
sqlalchemy 
python-decouple
mysql-connector-python
xhtml2pdf
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
Pillow
pytesseract
opencv-python
mysqlclient
datefinder
docraptor
